requirements.txt
python-dotenv>=1.0.1
pydantic>=2.6.0
pydantic-settings>=2.2.1

numpy>=1.26.0,<2.0.0
tqdm>=4.66.0

torch>=2.1.0,<2.6.0
sentence-transformers==2.7.0
transformers==4.41.2
huggingface-hub>=0.21.0,<1.0.0
faiss-cpu>=1.8.0,<2.0.0

# Optional: openai-whisper backend can be used if installed separately.
# openai-whisper>=20231117

soundfile>=0.12.1
faster-whisper>=1.0.3
